requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
css==0.1
Django==6.0.1
django-ckeditor==6.7.3
django-js-asset==3.1.2
etc==0.0.7
fanstatic==1.6
html5==0.0.9
idna==3.11
iso8601==2.1.0
javascript==1!1.2.6
js==1.0
legacy-cgi==2.6.4
mysql-connector==2.2.9
mysql-connector-python==9.5.0
netplan==0.3.3
networking==0.0.1
optional-django==0.3.0
packaging==26.0
pillow==12.1.1
postgres==4.0
psycopg2-binary==2.9.11
psycopg2-pool==1.2
PyYAML==6.0.3
react==4.3.0
requests==2.32.5
security==1.3.1
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
webencodings==0.5.1
WebOb==1.8.9
